e2e node: Verify pod worker is not blocked under frequent events
In kubelet, a pod worker gets the latest pod status from the cache
that is being updated by the PLEG every time the worker syncs a pod.
If another event to be handled happens while a pod worker is syncing a
pod, the worker tries to the newer pod status from the cache. If
EventedPLEG is enabled, the pod worker gets blocked at the cache much
longer than GenericPLEG. In order to verify this issue was fixed, this
change adds e2e tests that frequently raise events a pod workers should
handle in a short time.